rename vitest plugin entrypoint
