+++ code/addons/test/package.json
@@ -29,24 +29,24 @@
       "import": "./dist/index.js",
       "require": "./dist/index.cjs"
     },
-    "./vite-plugin": {
-      "types": "./dist/plugin/index.d.ts",
-      "import": "./dist/plugin/index.js",
-      "require": "./dist/plugin/index.cjs"
+    "./vitest-plugin": {
+      "types": "./dist/vitest-plugin/index.d.ts",
+      "import": "./dist/vitest-plugin/index.js",
+      "require": "./dist/vitest-plugin/index.cjs"
     },
     "./internal/global-setup": {
-      "types": "./dist/plugin/global-setup.d.ts",
-      "import": "./dist/plugin/global-setup.js",
-      "require": "./dist/plugin/global-setup.cjs"
+      "types": "./dist/vitest-plugin/global-setup.d.ts",
+      "import": "./dist/vitest-plugin/global-setup.js",
+      "require": "./dist/vitest-plugin/global-setup.cjs"
     },
     "./internal/setup-file": {
-      "types": "./dist/plugin/setup-file.d.ts",
-      "import": "./dist/plugin/setup-file.js"
+      "types": "./dist/vitest-plugin/setup-file.d.ts",
+      "import": "./dist/vitest-plugin/setup-file.js"
     },
     "./internal/test-utils": {
-      "types": "./dist/plugin/test-utils.d.ts",
-      "import": "./dist/plugin/test-utils.js",
-      "require": "./dist/plugin/test-utils.cjs"
+      "types": "./dist/vitest-plugin/test-utils.d.ts",
+      "import": "./dist/vitest-plugin/test-utils.js",
+      "require": "./dist/vitest-plugin/test-utils.cjs"
     },
     "./manager": "./dist/manager.js",
     "./preset": "./dist/preset.cjs",
@@ -93,16 +93,16 @@
   "bundler": {
     "exportEntries": [
       "./src/index.ts",
-      "./src/plugin/test-utils.ts",
-      "./src/plugin/setup-file.ts"
+      "./src/vitest-plugin/test-utils.ts",
+      "./src/vitest-plugin/setup-file.ts"
     ],
     "managerEntries": [
       "./src/manager.tsx"
     ],
     "nodeEntries": [
       "./src/preset.ts",
-      "./src/plugin/index.ts",
-      "./src/plugin/global-setup.ts",
+      "./src/vitest-plugin/index.ts",
+      "./src/vitest-plugin/global-setup.ts",
       "./src/postinstall.ts"
     ]
   }
